requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
datasets==4.1.0
diffusers==0.35.1
dill==0.4.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.34.6
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.5
numpy==2.3.3
packaging==25.0
pandas==2.3.2
peft==0.17.1
pillow==11.3.0
propcache==0.3.2
psutil==7.0.0
pyarrow==21.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
